requirements.txt
weasyprint==61.2
Jinja2==3.1.3
qrcode[pil]==7.4.2
Pillow==10.2.0
babel==2.14.0
num2words==0.5.13
python-dateutil==2.8.2
